+++ src/resources/views/master-product/_variants.blade.php
@@ -1,7 +1,7 @@
 <x-appshell::card>
     <x-slot:title>{{ __('Variants') }}</x-slot:title>
 
-    @foreach($product->variants as $variant)
+    @foreach($product->variants->sortByDesc('priority') as $variant)
         <a href="{{ route('vanilo.admin.master_product_variant.edit', [$product, $variant]) }}" class="badge bg-secondary mb-1">
             @if ($variant->hasImage())
                 <img src="{{ $variant->getThumbnailUrl() }}" style="width: 24px; border-radius: 50%" />